Added nudge functions '<' and '>'
THESE ARE UNTESTED, LIKE '[' AND ']'. Small scale tests are needed before testing with the carousel.
Nudges '<' and '>' for ABC and CBA rotation.
These cause accel() then deccel() without intermediate rotation, and are for testing.
That does not mean this is the minimum amount the carousel can be turned, since a custom function could be built to move it even less in one step.